Adiciona DTOs para criação de itens e usuários com validações

+++ backend/src/main/java/com/edvaldo/perdidos_achados/models/dto/item/requeste/ItemCreateDTO.java
@@ -23,6 +23,6 @@ public class ItemCreateDTO {
      private String localRef;
 
     @NotNull(message = "Id do usuário autenticado é necessário")
-    private Integer usuarioId;
+    private Long usuarioId;
 
 }
